rename wf_data to wf_models and remove commented old things
