+++ beeflow/common/wf_models.py
@@ -11,25 +11,6 @@ from typing import Optional, List, Dict
 
 from beeflow.common.container_path import convert_path
 
-'''
-# Workflow input parameter class
-InputParameter = namedtuple("InputParameter", ["id", "type", "value"])
-# Workflow output parameter class
-OutputParameter = namedtuple("OutputParameter", ["id", "type", "value", "source"])
-# Step input class
-StepInput = namedtuple("StepInput", ["id", "type", "value", "default", "source", "prefix",
-                                     "position", "value_from"])
-# Step output class
-StepOutput = namedtuple("StepOutput", ["id", "type", "value", "glob"])
-
-# CWL requirement class
-Requirement = namedtuple("Requirement", ["class_", "params"])
-# CWL hint class
-Hint = namedtuple("Hint", ["class_", "params"])
-# Task state update, usually sent from the task manager
-TaskStateUpdate = namedtuple("TaskStateUpdate", ["wf_id", "task_id", "job_state",
-                                                 "task_info", "output", "metadata"])
-'''
                                                  
 class InputParameter(BaseModel):
     """Pydantic model for InputParameter."""
